10-6-23
1. Tambah page ServiceList, Form, dan scs
2. Bikin api untuk edit new delete

+++ app/Helpers/updateImage.php
@@ -4,17 +4,19 @@ use App\Models\AboutUsModel;
 use Illuminate\Support\Facades\File;
 use Spatie\ImageOptimizer\OptimizerChainFactory;
 
-function SaveImage($file, $name, $imageNameFromDb)
+function SaveImage($file, $name, $imageNameFromDb = null)
 {
     $extension = '.'.$file->extension();
     $optimizerChain = OptimizerChainFactory::create();
     $optimizerChain->optimize($file);
-
-    $destinationPath = \base_path()."/public/images";
-    if(File::exists($destinationPath.'/'.$imageNameFromDb)){
-        File::delete($destinationPath.'/'.$imageNameFromDb);
+    
+    if(!empty($imageNameFromDb)){
+        $destinationPath = \base_path()."/public/images";
+        if(File::exists($destinationPath.'/'.$imageNameFromDb)){
+            File::delete($destinationPath.'/'.$imageNameFromDb);
+        }
     }
-
+   
     $imageName = $name.$extension;
     $file->move(\base_path()."/public/images", $imageName);
